Adjust generated types for 'unsigned char' conversion
TTypeKind.Char_U and TTypeKind.Char_S relate to type used if char is alias
for 'unsigned char' or 'signed char'.  For real 'unsigned char' and 'signed char'
values of TTypeKind.UChar and TTypeKind.SChar are used.  So in order to control
what is generated for 'unsigned char' conversion for TTypeKind.UChar should
be modified.

In the end probably there should be 3 settings: for 'char', 'unsigned char'
and 'signed char' separately. For now 'char' and 'signed char' will be
treated the same as this is most common on x86/x64 both on Linux and Windows.

Ref. http://lists.llvm.org/pipermail/cfe-dev/2012-February/019902.html
        https://www.linuxtopia.org/online_books/an_introduction_to_gcc/gccintro_71.html